pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clicktrack-player"
version = "0.1.0"
description = "A Raspberry Pi click track player for live musical performances"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Click Track Player Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Sound/Audio",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "Flask>=3.0.0",
    "Flask-SocketIO>=5.3.5",
    "Flask-CORS>=4.0.0",
    "python-rtmidi>=1.5.8",
    "PyAudio>=0.2.14",
    "simpleaudio>=1.0.4",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "hypothesis>=6.92.1",
    "pytest-cov>=4.1.0",
]

[project.scripts]
clicktrack = "clicktrack.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=clicktrack --cov-report=term-missing"
